\name{print.gcrq}
\alias{print.gcrq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print method for the gcrq class }
\description{
  Printing the most important feautures of a gcrq model.
}
\usage{
\method{print}{gcrq}(x, digits = max(3, getOption("digits") - 3), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object of class \code{gcrq} }
  \item{digits}{ number of digits to be printed }
  \item{\dots}{ arguments passed to other functions }
}
\author{ Vito M.R. Muggeo }

\seealso{ \code{\link{summary.gcrq}}}

\keyword{ models }
