\name{SiChildren}
\alias{SiChildren}
\docType{data}
\title{
Age, height and weight in a sample of Italian children
}
\description{
Age, height and weight in a sample of 1424 Italian children born in Sicily in the eighties
}
\usage{data("SiChildren")}
\format{
  A data frame with 1424 observations on the following 3 variables.
  \describe{
    \item{\code{age}}{age in years}
    \item{\code{height}}{child height (in centimeter) }
    \item{\code{weight}}{child weight (in kilo)}
  }
}
\details{
Data refer on the usual antropometric measures of Italian boys born in Sicily in the first years of 80s. Data have been kindly provided by prof M. Chiodi
}
\source{
Gattuccio F., and Pirronello S., and Chiodi M (1988)
Possibilita' di identificazione di tipologie evolutive del periodo puberale: proposta di una metodica pr finalita' predittive, \emph{Rivista di pediatria preventiva e sociale nipiologia}, 189-199
}
%\references{
%%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(SiChildren)
## see the package vignette for an example using such dataset
}
\keyword{datasets}
